chat: handle thread title and icon generation in a single call
